refactor: capitalized SearchForm because it was bugging me
